@@ -8,9 +8,10 @@ description: Demonstrates how to use the flutter_image_compress plugin.
 # build by specifying --build-name and --build-number, respectively.
 # Read more about versioning at semver.org.
 version: 1.0.0+1
+publish_to: none
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
@@ -18,12 +19,12 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
   flutter_image_compress:
     path: ../
-    
-  path_provider: ^0.4.1    
+
+  path_provider: ^2.0.0
 
 dev_dependencies:
   flutter_test:
@@ -34,7 +35,6 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -42,10 +42,9 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-   - img/
+    - img/
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
-
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
