@@ -43,10 +43,10 @@ class _MyAppState extends State<MyApp> {
     final result =
         await FlutterImageCompress.compressWithList(data.buffer.asUint8List());
 
-    print("after = ${result?.length ?? 0}");
+    print("after = ${result.length}");
   }
 
-  ImageProvider provider;
+  ImageProvider? provider;
 
   @override
   Widget build(BuildContext context) {
@@ -65,47 +65,47 @@ class _MyAppState extends State<MyApp> {
               ),
               aspectRatio: 1 / 1,
             ),
-            FlatButton(
+            TextButton(
               child: Text('CompressFile and rotate 180'),
               onPressed: _testCompressFile,
             ),
-            FlatButton(
+            TextButton(
               child: Text('CompressAndGetFile and rotate 90'),
               onPressed: getFileImage,
             ),
-            FlatButton(
+            TextButton(
               child: Text('CompressAsset and rotate 135'),
               onPressed: () => testCompressAsset("img/img.jpg"),
             ),
-            FlatButton(
+            TextButton(
               child: Text('CompressList and rotate 270'),
               onPressed: compressListExample,
             ),
-            FlatButton(
+            TextButton(
               child: Text('test compress auto angle'),
               onPressed: _compressAssetAndAutoRotate,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Test png '),
               onPressed: _compressPngImage,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Format transparent PNG'),
               onPressed: _compressTransPNG,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Restore transparent PNG'),
               onPressed: _restoreTransPNG,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Keep exif image'),
               onPressed: _compressImageAndKeepExif,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Convert to heic format and print the file url'),
               onPressed: _compressHeicExample,
             ),
-            FlatButton(
+            TextButton(
               child: Text('Convert to webp format, Just support android'),
               onPressed: _compressAndroidWebpExample,
             ),
@@ -138,6 +138,9 @@ class _MyAppState extends State<MyApp> {
     file.writeAsBytesSync(data.buffer.asUint8List());
 
     final result = await testCompressFile(file);
+
+    if (result == null) return;
+
     ImageProvider provider = MemoryImage(result);
     this.provider = provider;
     setState(() {});
@@ -182,11 +185,15 @@ class _MyAppState extends State<MyApp> {
     final targetPath = dir.absolute.path + "/temp.jpg";
     final imgFile = await testCompressAndGetFile(file, targetPath);
 
+    if (imgFile == null) {
+      return;
+    }
+
     provider = FileImage(imgFile);
     setState(() {});
   }
 
-  Future<Uint8List> testCompressFile(File file) async {
+  Future<Uint8List?> testCompressFile(File file) async {
     print("testCompressFile");
     final result = await FlutterImageCompress.compressWithFile(
       file.absolute.path,
@@ -196,11 +203,11 @@ class _MyAppState extends State<MyApp> {
       rotate: 180,
     );
     print(file.lengthSync());
-    print(result.length);
+    print(result?.length);
     return result;
   }
 
-  Future<File> testCompressAndGetFile(File file, String targetPath) async {
+  Future<File?> testCompressAndGetFile(File file, String targetPath) async {
     print("testCompressAndGetFile");
     final result = await FlutterImageCompress.compressAndGetFile(
       file.absolute.path,
@@ -212,7 +219,7 @@ class _MyAppState extends State<MyApp> {
     );
 
     print(file.lengthSync());
-    print(result.lengthSync());
+    print(result?.lengthSync());
 
     return result;
   }
@@ -226,6 +233,8 @@ class _MyAppState extends State<MyApp> {
       quality: 96,
       rotate: 135,
     );
+
+    if (list == null) return;
 
     this.provider = MemoryImage(Uint8List.fromList(list));
     setState(() {});
@@ -267,6 +276,9 @@ class _MyAppState extends State<MyApp> {
       quality: 95,
       // autoCorrectionAngle: false,
     );
+
+    if (result == null) return;
+
     final u8list = Uint8List.fromList(result);
     this.provider = MemoryImage(u8list);
     setState(() {});
@@ -278,6 +290,8 @@ class _MyAppState extends State<MyApp> {
       minWidth: 300,
       minHeight: 500,
     );
+
+    if (result == null) return;
 
     final u8list = Uint8List.fromList(result);
     this.provider = MemoryImage(u8list);
@@ -313,6 +327,8 @@ class _MyAppState extends State<MyApp> {
       keepExif: true,
     );
 
+    if (result == null) return;
+
     this.provider = MemoryImage(Uint8List.fromList(result));
     setState(() {});
 
@@ -335,6 +351,9 @@ class _MyAppState extends State<MyApp> {
       format: CompressFormat.heic,
       quality: 90,
     );
+
+    if (result == null) return;
+
     print("Compress heic success.");
     logger.logTime();
     print("src, path = $srcPath length = ${File(srcPath).lengthSync()}");
@@ -360,6 +379,9 @@ class _MyAppState extends State<MyApp> {
       minWidth: 800,
       quality: quality,
     );
+
+    if (result == null) return;
+
     print("Compress webp success.");
     logger.logTime();
     print("src, path = $srcPath length = ${File(srcPath).lengthSync()}");
@@ -377,10 +399,10 @@ Future<Uint8List> getAssetImageUint8List(String key) async {
 }
 
 double calcScale({
-  double srcWidth,
-  double srcHeight,
-  double minWidth,
-  double minHeight,
+  required double srcWidth,
+  required double srcHeight,
+  required double minWidth,
+  required double minHeight,
 }) {
   final scaleW = srcWidth / minWidth;
   final scaleH = srcHeight / minHeight;
