@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter_image_compress/flutter_image_compress.dart';
 
 class TryCatchExample {
-  Future<List<int>> compressAndTryCatch(String path) async {
-    List<int> result;
+  Future<List<int>?> compressAndTryCatch(String path) async {
+    List<int>? result;
     try {
       result = await FlutterImageCompress.compressWithFile(path,
           format: CompressFormat.heic);
