@@ -1,13 +1,6 @@
 package com.example.flutterimagecompressexample
 
-import android.os.Bundle
+import io.flutter.embedding.android.FlutterActivity
 
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
-
-class MainActivity(): FlutterActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    GeneratedPluginRegistrant.registerWith(this)
-  }
+class MainActivity() : FlutterActivity() {
 }
